fix(client): honor _-prefix in eslint + expires_at u64->f64 wire-type drift
System-check findings: eslint config now ignores ^_ intentionally-unused
bindings (was failing on voice-ask-wiring _event/_speak); SessionCompleteResponse.expires_at
typed f64 to match the brain's float wire type (u64 fails serde on a fractional timestamp).

+++ client/src-tauri/src/gateway.rs
@@ -36,7 +36,9 @@ pub(crate) struct SessionCompleteRequest {
 #[derive(Debug, Deserialize, Serialize)]
 pub(crate) struct SessionCompleteResponse {
     session_token: String,
-    expires_at: u64,
+    // Brain serializes this as a float Unix timestamp (api_app SessionCompleteResponse.expires_at: float);
+    // u64 would fail serde deserialization on any fractional value. Deserialize-only — never used in arithmetic.
+    expires_at: f64,
 }
 
 #[derive(Debug, Deserialize, Serialize)]
